vcf_writer added :information_desk_person:
